refactor(chat): streamline chat logic and remove redundant code
- Restructured ChatInfo to simplify chat management
- Merged handleEditorConfirmation and processSingleModeInput into processEditorInput
- Removed unused showInfo, showHelp, and handleCommand functions
- Updated chat flow to use new processEditorInput method
- Added resetInputState function to reset chat input state
- Updated version to v1.12.1
